Update Linux Version
Now uses shared memory, with the added benefit of being faster and not requiring sudo permission.
However, due to Dolphin immediately unlinking the shared memory after its created on Linux, the library now has to run before a game is started while constantly attempting to access the shared memory object. A bug with Python also causes the object to be deleted after the process ends.
The only superficial effect of this update on Linux is that a warning is posted after the library finishes its test.